Parsing and Validating Credit Cards
This contains the class ECCreditCard. The class contains two readonly
properties - cardIssuer & isValid. The class also has a default
initializer -initWithCreditCardNumber: which takes credit card number
as NSString instance.

The CreditCardValidatorTests is the test class which has various unit
tests to test ECCreditCard. The unit tests have a code coverage of 86%
of ECCreditCard.